Config: change CMA_SIZE to 0
This patch disables CMA to let the system use the default swiotlb
DMA allocation. It is observed certain system configuration such as
1 LSI and 2 Mellanox cards failed to initialize if they failed to allocate
CMA buffer. But they would function properly if there is no CMA buffer
to start with. This issue is fixed from upstream 4.14 kernel. For more
recent kernel, the patch will not be needed.

Change-Id: I07cff5d4c3f46aa15ee610e8991659ce80583634
Signed-off-by: Khuong Dinh <kdinh@apm.com>